test(process): own shell fixture process groups
Move the existing daemon fixture owner into a shared panic-safe test helper and use it for every raw shell-grandchild liveness fixture in #3004.
